Consolidate public reservation feature tests into one suite (#120)
Closes #28

Refs PRD-002. Introduces tests/Feature/PublicReservationTest.php
with the 10 acceptance-criteria tests (form rendering, valid
submission persistence, event dispatch, past-date rejection, party
size cap, malformed email rejection, silent honeypot, rate limit,
404 unknown slug, UTC persistence) plus the tonality greeting case
and the Inertia-friendly throttle case that were already earning
their keep.

Removes the now-absorbed PublicReservationRoutesTest and
PublicReservationStoreTest entirely. Trims duplicated behaviours
out of StoreReservationRequestTest and renames the retained field
tests to reflect that their sole remaining role is to pin the
German error messages.
